Private Bug fixed / Authurize Improved

+++ src/Presentation/Virgol.School/Controllers/School/SchoolController.cs
@@ -113,7 +113,7 @@ namespace lms_with_moodle.Controllers
 
                 SchoolModel schoolModel = appDbContext.Schools.Where(x => x.Id == schoolId).FirstOrDefault();
 
-                UserModel managerInfo = appDbContext.Users.Where(x => x.SchoolId == schoolModel.Id && x.UserType == Roles.Manager).FirstOrDefault();
+                UserModel managerInfo = appDbContext.Users.Where(x => x.Id == schoolModel.ManagerId).FirstOrDefault();
                 ManagerDetail managerDetail = new ManagerDetail();
                 if(managerInfo != null)
                 {
@@ -294,7 +294,7 @@ namespace lms_with_moodle.Controllers
                 manager.UserName = inputData.MelliCode;
                 manager.PhoneNumber = inputData.managerPhoneNumber;
                 manager.SchoolId = schoolResult.Id;
-                manager.UserType = Roles.Manager;
+                //manager.UserType = Roles.Manager;
                 manager.ConfirmedAcc = true;
 
                 string password = RandomPassword.GeneratePassword(true , true , true , 8);
@@ -413,7 +413,7 @@ namespace lms_with_moodle.Controllers
 
                 if(removeCat)
                 {
-                    UserModel manager = appDbContext.Users.Where(x => x.SchoolId == schoolId && x.UserType == Roles.Manager).FirstOrDefault();
+                    UserModel manager = appDbContext.Users.Where(x => x.Id == school.ManagerId).FirstOrDefault();
                     if(manager != null)
                     {
                         // manager.SchoolId = -1;
@@ -427,21 +427,25 @@ namespace lms_with_moodle.Controllers
                     appDbContext.School_Grades.RemoveRange(appDbContext.School_Grades.Where(x => x.School_Id == school.Id).ToList());
                     appDbContext.School_Classes.RemoveRange(appDbContext.School_Classes.Where(x => x.School_Id == school.Id).ToList());
 
-                    List<UserModel> students = appDbContext.Users.Where(x => x.UserType == Roles.Student && x.SchoolId == school.Id).ToList();
+                    List<UserModel> students = appDbContext.Users.Where(x => x.SchoolId == school.Id).ToList();
 
                     foreach (var student in students)
                     {
                         try
                         {
-                            await UserService.DeleteUser(student);
-                            
-                            School_studentClass stdClass = appDbContext.School_StudentClasses.Where(x => x.UserId == student.Id).FirstOrDefault();
-                            StudentDetail stdDetail = appDbContext.StudentDetails.Where(x => x.UserId == student.Id).FirstOrDefault();
-                            ParticipantInfo participant = appDbContext.ParticipantInfos.Where(x => x.UserId == student.Id).FirstOrDefault();
+                            //Second operator just for old User
+                            if(UserService.HasRole(student , Roles.Student) || UserService.HasRole(student , Roles.User , true))
+                            {
+                                await UserService.DeleteUser(student);
+                                
+                                School_studentClass stdClass = appDbContext.School_StudentClasses.Where(x => x.UserId == student.Id).FirstOrDefault();
+                                StudentDetail stdDetail = appDbContext.StudentDetails.Where(x => x.UserId == student.Id).FirstOrDefault();
+                                ParticipantInfo participant = appDbContext.ParticipantInfos.Where(x => x.UserId == student.Id).FirstOrDefault();
 
-                            appDbContext.ParticipantInfos.Remove(participant);
-                            appDbContext.StudentDetails.Remove(stdDetail);
-                            appDbContext.School_StudentClasses.Remove(stdClass);
+                                appDbContext.ParticipantInfos.Remove(participant);
+                                appDbContext.StudentDetails.Remove(stdDetail);
+                                appDbContext.School_StudentClasses.Remove(stdClass);
+                            }
                             
                         }catch{}
                     }
@@ -1024,7 +1028,7 @@ namespace lms_with_moodle.Controllers
                             manager.MelliCode = schoolData.MelliCode;
                             manager.UserName = schoolData.MelliCode;
                             manager.SchoolId = schoolResult.Id;
-                            manager.UserType = Roles.Manager;
+                            //manager.UserType = Roles.Manager;
                             manager.ConfirmedAcc = true;
 
                             //string password = RandomPassword.GeneratePassword(true , true , true , 10);
@@ -1040,7 +1044,7 @@ namespace lms_with_moodle.Controllers
                                 int userId = userManager.FindByNameAsync(manager.UserName).Result.Id;
                                 manager.Id = userId;
 
-                                bool ldapUser = ldap.AddUserToLDAP(manager , password);
+                                bool ldapUser = ldap.AddUserToLDAP(manager , Roles.Manager , password);
 
                                 bool userToMoodle = false;
                                 if(ldapUser)
